refactor(profiling): promote profiling to module, extract uploader
